[BE] refactor: controller 정리
private, public router 구분이 없어지면서,
contorller 또한 구분이 필요 없게 되었다.
