+++ backend/routes/library/index.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
-import libraryController from "../../controllers/public/library";
-import userLikedItems from "../../controllers/public/userLikedItems";
+import libraryController from "../../controllers/library";
+import userLikedItems from "../../controllers/userLikedItems";
 
 const router = Router();
 router.get("/test", userLikedItems);
