+++ backend/routes/search/index.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import searchController from "../../controllers/public/search";
+import searchController from "../../controllers/search";
 
 const router = Router();
 router.get("/", searchController.getAll);
