+++ backend/routes/tracks/index.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import trackController from "../../controllers/public/tracks";
+import trackController from "../../controllers/tracks";
 
 const router = Router();
 router.post("/like", trackController.like);
